feat/#7 - 일정 정보 조회, 수정, 삭제 시 세션 정보 사용하도록 변경
- 일정의 수정, 삭제 등은 해당 일정을 작성한 회원만 가능하도록 했다.

+++ src/main/java/xyz/ncookie/sma/schedule/service/ScheduleService.java
@@ -14,6 +14,8 @@ import xyz.ncookie.sma.schedule.dto.response.ScheduleResponseDto;
 import xyz.ncookie.sma.schedule.entity.Schedule;
 import xyz.ncookie.sma.schedule.repository.ScheduleRepository;
 
+import java.util.Objects;
+
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -24,9 +26,9 @@ public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
 
     @Transactional
-    public ScheduleResponseDto saveSchedule(ScheduleSaveRequestDto dto) {
+    public ScheduleResponseDto saveSchedule(Long memberId, ScheduleSaveRequestDto dto) {
 
-        Member member = memberRetrievalService.findById(dto.memberId());
+        Member member = memberRetrievalService.findById(memberId);
 
         Schedule savedSchedule = scheduleRepository.save(
                 Schedule.of(
@@ -46,9 +48,11 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long scheduleId, ScheduleUpdateRequestDto dto) {
+    public ScheduleResponseDto updateSchedule(Long scheduleId, Long memberId, ScheduleUpdateRequestDto dto) {
 
         Schedule findSchedule = findById(scheduleId);
+
+        verifyScheduleOwner(findSchedule, memberId);
 
         findSchedule.updateSchedule(dto.title(), dto.contents());
 
@@ -57,9 +61,12 @@ public class ScheduleService {
     }
 
     @Transactional
-    public void deleteScheduleById(Long scheduleId) {
+    public void deleteScheduleById(Long scheduleId, Long memberId) {
 
         Schedule findSchedule = findById(scheduleId);
+
+        verifyScheduleOwner(findSchedule, memberId);
+
         scheduleRepository.delete(findSchedule);
     }
 
@@ -67,6 +74,15 @@ public class ScheduleService {
 
         return scheduleRepository.findById(scheduleId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "존재하지 않는 ID입니다. = " + scheduleId));
+    }
+
+    // 일정 수정 또는 삭제 시 해당 일정을 작성한 회원인지 검증
+    private void verifyScheduleOwner(Schedule findSchedule, Long sessionMemberId) {
+
+        Long findMemberId = findSchedule.getMember().getId();
+        if (!Objects.equals(sessionMemberId, findMemberId)) {
+            throw new RuntimeException("본인이 작성한 일정만 수정 또는 삭제할 수 있습니다. 일정 ID, 회원 ID: " + findSchedule.getId() + ", " + findMemberId);
+        }
     }
 
 }
